move /Wallet/DelegatedValidator to home tab: Validators
