+++ src/components/DelegatedValidator/DelegatedClaim/ClaimCardList/ClaimCardList.module.scss
