+++ src/components/DelegatedValidator/DelegatedClaim/ClaimCardList/index.js
