+++ src/components/DelegatedValidator/DelegatedClaim/ClaimTable/ClaimRwAllBtn/index.js
