+++ src/components/DelegatedValidator/DelegatedClaim/ClaimTable/ClaimTable.js
