+++ src/components/DelegatedValidator/DelegatedClaim/ClaimTable/ClaimTable.module.scss
