+++ src/components/DelegatedValidator/DelegatedClaim/ClaimTable/ClaimTableSkeleton.js
