+++ src/components/DelegatedValidator/DelegatedClaim/ClaimTable/index.js
