+++ src/components/DelegatedValidator/DelegatedClaim/DelegatedClaim.module.scss
