+++ src/components/DelegatedValidator/DelegatedClaim/DelegatedTable/Claim/Claim.module.scss
