+++ src/components/DelegatedValidator/DelegatedClaim/DelegatedTable/Claim/index.js
