+++ src/components/DelegatedValidator/DelegatedValidator.module.scss
@@ -2,6 +2,7 @@
 	padding: 0;
 	background: transparent;
 	box-shadow: none;
+	margin-bottom: 40px;
 
 	@media screen and (min-width: 1280px) {
 		padding: 20px;
