+++ src/components/DelegatedValidator/DelegatedWithdraw/DelegatedWithdraw.js
@@ -6,10 +6,10 @@ import {useTheme} from "@material-ui/core/styles";
 import useMediaQuery from "@material-ui/core/useMediaQuery";
 import consts from "src/constants/consts";
 import NoResult from "src/components/common/NoResult";
-import WithdrawTable from "src/components/Wallet/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTable";
-import WithdrawTableSkeleton from "src/components/Wallet/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTableSkeleton";
-import WithdrawCardList from "src/components/Wallet/DelegatedValidator/DelegatedWithdraw/WithdrawCardList/WithdrawCardList";
-import WithdrawCardListSkeleton from "src/components/Wallet/DelegatedValidator/DelegatedWithdraw/WithdrawCardList/WithdrawCardListSkeleton";
+import WithdrawTable from "src/components/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTable";
+import WithdrawTableSkeleton from "src/components/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTableSkeleton";
+import WithdrawCardList from "src/components/DelegatedValidator/DelegatedWithdraw/WithdrawCardList/WithdrawCardList";
+import WithdrawCardListSkeleton from "src/components/DelegatedValidator/DelegatedWithdraw/WithdrawCardList/WithdrawCardListSkeleton";
 import arrowIcon from "src/assets/wallet/arrow.svg";
 import styles from "./DelegatedWithdraw.module.scss";
 
@@ -46,10 +46,6 @@ export default function({setActiveTab, address}) {
 					<div className={cx("title")}>Withdraw tokens</div>
 					<p className={cx("note")}>Please be aware that you have to wait 14 days to complete unbonding your funds from validators.</p>
 				</div>
-				<button className={cx("button")} onClick={() => setActiveTab(0)}>
-					Claim Reward
-					<img className={cx("button-icon")} src={arrowIcon} />
-				</button>
 			</div>
 			{tableSection}
 		</div>
