+++ src/components/DelegatedValidator/DelegatedWithdraw/DelegatedWithdraw.module.scss
