+++ src/components/DelegatedValidator/DelegatedWithdraw/WithdrawCardList/WithdrawCardList.module.scss
