+++ src/components/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawBtn/WithdrawBtn.module.scss
