+++ src/components/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTable.module.scss
