+++ src/components/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTableSkeleton.js
@@ -3,7 +3,7 @@ import Skeleton from "@material-ui/lab/Skeleton";
 import classNames from "classnames/bind";
 import {tableThemes} from "src/constants/tableThemes";
 import ThemedTable from "src/components/common/ThemedTable";
-import {getHeaderRow} from "src/components/Wallet/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTable";
+import {getHeaderRow} from "src/components/DelegatedValidator/DelegatedWithdraw/WithdrawTable/WithdrawTable";
 import styles from "./WithdrawTable.module.scss";
 
 const WithdrawTableSkeleton = memo(({rows = 5}) => {
