+++ src/components/DelegatedValidator/index.js
@@ -1,0 +1,3 @@
+import DelegatedValidator from "./DelegatedValidator1";
+
+export default DelegatedValidator;
