--- src/components/Wallet/DelegatedValidator/index.js
@@ -1,3 +1,0 @@
-import DelegatedValidator from "./DelegatedValidator";
-
-export default DelegatedValidator;
