+++ src/components/Wallet/Tabs/Tabs.js
@@ -4,7 +4,6 @@ import cn from "classnames/bind";
 import {useGet} from "restful-react";
 import ContactIcon from "src/icons/Tabs/ProposalsTabIcon";
 import ValidatorsIcon from "src/icons/Tabs/ValidatorsTabIcon";
-import DelegatedIcon from "src/icons/Validators/ValidatorsIcon";
 import TransactionsIcon from "src/icons/Tabs/TransactionsTabIcon";
 import {getListCwToken} from "src/lib/api";
 import styles from "./Tabs.module.scss";
@@ -33,10 +32,6 @@ export default function({activeTab, setActiveTab, isBecomeValidator, address}) {
 					<div className={cx("tab-text")}>Your Delegators</div>
 				</div>
 			)}
-			<div className={cx("tab", activeTab === 2 ? "active" : "")} onClick={() => setActiveTab(2)}>
-				<DelegatedIcon className={cx("tab-icon")} />
-				<div className={cx("tab-text")}>Delegated Validator</div>
-			</div>
 			<div className={cx("tab", activeTab === 4 ? "active" : "")} onClick={() => setActiveTab(4)}>
 				<ContactIcon className={cx("tab-icon")} />
 				<div className={cx("tab-text")}>Contact</div>
